Mean Decrease Impurity
MDI is specific to tree-based classifiers, like RF. At each node of decision tree, the split feature is selected from a random subset of all features.  The score of each feature is chosen to be the the number of split using this feature over all splits. 

MDI doesn't account OOS performance. Thus it is fast.